Update firmware(9) image names.
Based on manu's comments in https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=251415#c5
resubmit this change, which previously was part of
https://github.com/FreeBSDDesktop/kms-firmware/pull/13
to be included in the main kms-firmware repo.

This change will adjust the firmware image name to the firmware
kernel module file name and with that will allow firmware(9)
as-is to auto-load the the kernel module without any extra logic
needed for firmare(9)/kernel linker.

Sposnored by:	The FreeBSD Foundation